fix datatype issue

+++ src/solve.py
@@ -22,7 +22,7 @@ args = parser.parse_args()
 
 import time
 minutes = args.timetest*60
-checkpoints = minutes / 10 #report on 10 second intervals
+checkpoints = int(minutes / 10) #report on 10 second intervals
 for c in range(0, checkpoints):
     print(f'{c} seconds have passed')
     time.sleep(10)
